+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this package will be documented in this file.
 
 ## [Unreleased]
+- Changed: ButtonMethodAttribute - new setting to draw button before or after inspector
 ! wip AttributeBase
 
 ## [1.4.1] - 2020-06-06